16.2: Fix direction
...but then the answer (432) is too low.

I'm very close. I just need to analyze the fork position. I need to pick
a direction where there isn't a wall, which might not be the direction
of the fork.

Alternatively (and more simply), I could just solve the maze from the
start to the fork and then continue from the fork to the end. This is
the much better solution. Then, I can just continue from exactly the
cell where I ended.